Added bdc_color_even.m
Used to get colormaps with specified number of elements in each
bin. E.g. bdc_color_even(z, [ 0.1, 0.5]) would put in color
transitions in the color map in such a way that 10% of the
elements in z are in the first bin 50% - 10% = 40% in the next
and the last 50% in the last.